# public/swagger-ui/swagger.json
{
    "swagger": "2.0",
    "info": {  
        "description": "Project to manage information about fields, tractors and processed fields",
        "version": "1.0.0",
        "title": "Manage Agriculture Field",
        "contact": {  
            "email": "deepakms001@yahoo.com"
        }
    },
    "host": "localhost:8000",
    "basePath": "",
...
    "schemes": [  
        "http"
    ],
    "paths": {  
...
        "/fields": {  
            "get": {  
                "tags": [  
                    "Fields"
                ],
                "summary": "Get all process fields requests",
                "description": "Get the list of process fields requests",
                "operationId": "/allFields",
                "produces": [  
                    "application/json"
                ],
                "parameters": [  
                    {  
                        "name": "token",
                        "in": "query",
                        "description": "Token obtained from authentication API ",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {  
                    "200": {  
                        "description": "successful operation",
                        "schema": {  
                            "type": "array",
                            "items": {  
                                "$ref": "#/definitions/Pet"
                            }
                        }
                    },
                    "400": {  
                        "description": "Provided token is expired",
                        "schema": {  
                            "type": "object",
                            "items": {
                                "status": "error",
                                "message": "Provided token is expired"
                            }
                        }
                    }
                }
            },
            "post": {  
                "tags": [  
                    "Fields"
                ],
                "summary": "Add a field",
                "description": "",
                "operationId": "addField",
                "consumes": [  
                    "application/x-www-form-urlencoded"
                ],
                "produces": [  
                    "application/json"
                ],
                "parameters": [  
                    {  
                        "name": "name",
                        "in": "formData",
                        "description": "Name of the field",
                        "required": true,
                        "type": "string"
                    }, {  
                        "name": "user_id",
                        "in": "formData",
                        "description": "Can only be set by admin. Customers need not set this value.",
                        "required": false,
                        "type": "string"
                    },
                    {  
                        "name": "token",
                        "in": "formData",
                        "description": "Token optained from auth API",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {  
                    "405": {  
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/fields/{fieldId}": {
            "delete": {  
                "tags": [  
                    "Fields"
                ],
                "summary": "Delete a field",
                "description": "",
                "operationId": "deleteField",
                "produces": [  
                    "application/json"
                ],
                "parameters": [  
                    {  
                        "name": "token",
                        "in": "query",
                        "description": "Token obtained from auth API",
                        "required": true,
                        "type": "string"
                    },
                    {  
                        "name": "tractorId",
                        "in": "path",
                        "description": "ID of field to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {  
                    "400": {  
                        "description": "Invalid ID supplied"
                    },
                    "404": {  
                        "description": "Pet not found"
                    }
                }
            }
        },
        "/process-fields": {  
            "get": {  
                "tags": [  
                    "Process Fields"
                ],
                "summary": "Get all fields",
                "description": "Get all fields list",
                "operationId": "/allProcessFields",
                "produces": [  
                    "application/json"
                ],
                "parameters": [  
                    {  
                        "name": "token",
                        "in": "query",
                        "description": "Token obtained from authentication API ",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {  
                    "200": {  
                        "description": "successful operation",
                        "schema": {  
                            "type": "array",
                            "items": {  
                                "$ref": "#/definitions/Pet"
                            }
                        }
                    },
                    "400": {  
                        "description": "Provided token is expired",
                        "schema": {  
                            "type": "object",
                            "items": {
                                "status": "error",
                                "message": "Provided token is expired"
                            }
                        }
                    }
                }
            },
            "post": {  
                "tags": [  
                    "Process Fields"
                ],
                "summary": "Add a process field request",
                "description": "",
                "operationId": "addProcessField",
                "consumes": [  
                    "application/x-www-form-urlencoded"
                ],
                "produces": [  
                    "application/json"
                ],
                "parameters": [  
                    {  
                        "name": "token",
                        "in": "formData",
                        "description": "Token obtained from auth API",
                        "required": true,
                        "type": "string"
                    },
                    {  
                        "name": "date",
                        "in": "formData",
                        "description": "Date of processing",
                        "required": true,
                        "type": "string",
                        "format": "date"
                    },
                    {  
                        "name": "field_id",
                        "in": "formData",
                        "description": "Updated field_id of process field request. Customers can only change if the own the field.",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {  
                        "name": "tractor_id",
                        "in": "formData",
                        "description": "Updated tractor_id of process field request.",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {  
                    "405": {  
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {  

        "token": {  
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    },
    "definitions": {  
...
        "Category": {  
            "type": "object",
            "properties": {  
                "id": {  
                    "type": "integer",
                    "format": "int64"
                },
                "name": {  
                    "type": "string"
                }
            },
            "xml": {  
                "name": "Category"
            }
        },
        "Tag": {  
            "type": "object",
            "properties": {  
                "id": {  
                    "type": "integer",
                    "format": "int64"
                },
                "name": {  
                    "type": "string"
                }
            },
            "xml": {  
                "name": "Tag"
            }
        },
        "Pet": {  
            "type": "object",
            "required": [  
                "name",
                "photoUrls"
            ],
            "properties": {  
                "id": {  
                    "type": "integer",
                    "format": "int64"
                },
                "category": {  
                    "$ref": "#/definitions/Category"
                },
                "name": {  
                    "type": "string",
                    "example": "doggie"
                },
                "photoUrls": {  
                    "type": "array",
                    "xml": {  
                        "name": "photoUrl",
                        "wrapped": true
                    },
                    "items": {  
                        "type": "string"
                    }
                },
                "tags": {  
                    "type": "array",
                    "xml": {  
                        "name": "tag",
                        "wrapped": true
                    },
                    "items": {  
                        "$ref": "#/definitions/Tag"
                    }
                },
                "status": {  
                    "type": "string",
                    "description": "pet status in the store",
                    "enum": [  
                        "available",
                        "pending",
                        "sold"
                    ]
                }
            },
            "xml": {  
                "name": "Pet"
            }
        },
...
    },
...
}